summarization-workflow/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
appdirs
arxiv
asgiref
asttokens
attrs
auth0-python
azure-common
azure-core
azure-search-documents
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
cohere
colorama
coloredlogs
comm
community
crewai
crewai-tools
cryptography
dataclasses-json
debugpy
decorator
Deprecated
deprecation
distro
dnspython
docker
docstring_parser
duckduckgo_search
durationpy
email_validator
embedchain
et_xmlfile
executing
faiss-cpu
fastapi
fastapi-cli
fastavro
feedparser
filelock
Flask
flask-cors
Flask-Limiter
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
gptcache
greenlet
groq
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
idna
importlib_metadata
importlib_resources
instructor
ipykernel
ipython
ipython_pygments_lexers
isodate
itsdangerous
jedi
Jinja2
jiter
joblib
json5
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kubernetes
lancedb
langchain
langchain-chroma
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-groq
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langsmith
limits
litellm
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mem0ai
mmh3
monotonic
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nodeenv
numpy
oauthlib
ollama
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
ordered-set
orjson
overrides
packaging
pandas
parso
pdfminer.six
pdfplumber
pillow
platformdirs
portalocker
posthog
primp
prompt_toolkit
propcache
protobuf
psutil
psycopg2-binary
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
PyJWT
pymongo
pypdf
pypdfium2
PyPika
pyproject_hooks
pyreadline3
pyright
pysbd
python-dateutil
python-dotenv
python-multipart
pytube
pytz
pyvis
#pywin32
PyYAML
pyzmq
qdrant-client
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rpds-py
rsa
safetensors
schema
scikit-learn
scipy
sentence-transformers
setuptools
sgmllib3k
shellingham
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
sympy
tabulate
tenacity
threadpoolctl
tiktoken
tokenizers
tomli
tomli_w
torch
tornado
tqdm
traitlets
transformers
typer
types-requests
typing-inspect
typing_extensions
tzdata
ujson
urllib3
uv
uvicorn
watchfiles
wcwidth
websocket-client
websockets
Werkzeug
wikipedia
wrapt
yarl
zipp
zstandard
